+++ src/irmin-pack/inode_intf.ml
@@ -47,6 +47,8 @@ module type S = sig
 
   val decode_bin :
     dict:(int -> string option) -> hash:(int64 -> key) -> string -> int -> int
+
+  val integrity_check_inodes : [ `Read ] t -> key -> (unit, string) result Lwt.t
 end
 
 (** Unstable internal API agnostic about the underlying storage. Use it only to
@@ -78,6 +80,9 @@ module type INTER = sig
     val stable : t -> bool
     val length : t -> int
     val index : depth:int -> step -> int
+
+    val integrity_check : t -> bool
+    (** Checks the integrity of an inode. *)
   end
 end